site-visit-report-app/backend/requirements.txt
# Core Web Framework
flask==3.0.2
flask-cors==4.0.0
gunicorn==23.0.0
python-dotenv==1.0.1

# Image Processing
Pillow==10.2.0

# HTTP Requests
requests==2.31.0

# PDF Generation
reportlab==4.0.9
fpdf==1.7.2
PyPDF2==3.0.1

# AI/ML Components (Compatible versions)
langchain==0.2.5
langchain-community==0.2.5
faiss-cpu==1.7.4
sentence-transformers==2.5.1

# Security
cryptography==44.0.1 